refactor(MNT-05): extract MemoryScenarioContextBuilder from MemoryBenchmarkRunner god class
MemoryBenchmarkRunner mixed orchestration with scenario authoring. Move the pure context-building statics
(BuildInterleavedHistory, InsertFactIntoSegment, BuildInterleavedHistoryNoSessions,
BuildContextPressureBlob, FormatTurnsAsTextBlob) into a dedicated internal MemoryScenarioContextBuilder
(the ScenarioContextBuilder the review names), shrinking the runner ~1183 to ~975 lines and redirecting
the call sites + one test. Behavior preserved (500 Memory tests + main suite green). The two constructors
already chain (not duplicated) and the per-run override race was already fixed (BUG-56); the deeper SRP
split (a JsonScenarioRunner for the 12 handlers + removing the defensive hardcoded fallbacks) remains
vNext as it is risky with no functional gain.
